ARQGRA-240: custom @FindBy moved from SPI to IMPL
It will be in IMPL till API module will be substracted. I also deleted
including of JQuery.js from staleelements.html as it threw exceptions with
HTMLUnit driver and it is not neccessary to be included.
