+++ graphene-webdriver/graphene-webdriver-impl/src/main/java/org/jboss/arquillian/graphene/enricher/findby/FindBy.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.arquillian.graphene.spi.annotations;
+package org.jboss.arquillian.graphene.enricher.findby;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
